requirements.txt
fastapi
uvicorn
langchain
langchain-text-splitters
chromadb
python-multipart
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-magic
python-magic-bin;platform_system=='Windows'
mysql-connector-python
pymysql
python-dotenv>=1.0.0


# 向量数据库
chromadb>=0.4.0

# 文档处理
pypdf>=3.0.0
docx2txt>=0.8
pdf2image>=1.16.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
rank_bm25>=0.1.0
jieba>=0.42.1

